- moved files to src - added src folder in the Dockerfile
