+++ README.md
@@ -53,6 +53,14 @@ spec:
                 - 'helloworld'
 ```
 
+To start it locally locally:
+
+```
+minikube start
+kubectl create -f ./cronjob.example.yml
+kubectl create job test-1 --from=cronjob/myjob
+```
+
 # Deployment of this example:
 ```yaml
 apiVersion: extensions/v1beta1
