+++ __tests__/etl.spec.js
@@ -1,5 +1,5 @@
-const s3 = require("../modules/s3")
-const kube = require("../modules/kube")
+const s3 = require("../src/s3")
+const kube = require("../src/kube")
 
 describe("etl", () => {
   describe("sqs_message_handler", () => {
@@ -11,20 +11,20 @@ describe("etl", () => {
     })
 
     it("should start the kube job if the manifest is good", async () => {
-      jest.mock("../modules/s3", () => ({
+      jest.mock("../src/s3", () => ({
         check_manifest: () => Promise.resolve(true)
       }))
-      const etl = require("../etl.js")
+      const etl = require("../src/etl.js")
 
       await etl.sqs_message_handler("test", done)
       expect(kube.start_kube_job).toHaveBeenCalledTimes(1)
     })
 
     it("should not start the kube job if the manifest is incorrect", async () => {
-      jest.mock("../modules/s3", () => ({
+      jest.mock("../src/s3", () => ({
         check_manifest: () => Promise.resolve(false)
       }))
-      const etl = require("../etl.js")
+      const etl = require("../src/etl.js")
 
       await etl.sqs_message_handler("test", done)
       expect(kube.start_kube_job).toHaveBeenCalledTimes(1)
