+++ src/etl.js
@@ -2,8 +2,8 @@
 
 const {BUCKET, ROLE, CRONJOB} = process.env
 
-const {check_manifest} = require("./modules/s3")
-const {start_kube_job} = require("./modules/kube")
+const {check_manifest} = require("./s3")
+const {start_kube_job} = require("./kube")
 
 const sqs_message_handler = async (message, done) => {
   if (await check_manifest(BUCKET)) await start_kube_job(ROLE, CRONJOB)
